04-1_realState/requirement.txt
Django==3.1
dj-rest-auth==1.1.1
django-allauth==0.42.0
django-debug-toolbar==2.2
django-extensions==3.0.5
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
django-bootstrap4==2.2.0
drf-yasg==1.17.1
lxml==4.5.2
tqdm==4.48.2
Markdown==3.2.2
whitenoise==5.2.0
gunicorn==20.0.4
Pillow==7.2.0